Modularize the service too
